Moving d1-api-test into its own file in the extensions dir
